MomentClosure.jl API

Model definition

As described in the tutorial on model initialisation, MomentClosure is fully compatible with reaction network models defined using Catalyst and stored as a ModelingToolkit.ReactionSystem. Nevertheless, we have implemented a limited extension of ModelingToolkit.ReactionSystem, called ReactionSystemMod, that allows us to consider systems containing reactions whose products can be independent geometrically distributed random variables. Note that such systems, one being the gene network example [LINK], are currently not supported by Catalyst. Moreover, ReactionSystemMod provides an alternative to Catalyst as the model is now constructed directly in terms of user-defined reaction propensity functions and the stoichiometric matrix, with system parameters and species defined using ModelingToolkit @parameters and @variables respectively. However, we stress that defining models using Catalyst is preferred whenever possible due to its much richer integration within the broader SciML ecosystem.

ModelingToolkit.@parameters — Macro

Define one or more known variables.

ModelingToolkit.@variables — Macro

Define one or more unknown variables.

@parameters t α σ(..) β[1:2]
@variables w(..) x(t) y() z(t, α, x)

expr = β₁* x + y^α + σ(3) * (z - t) - β₂ * w(t - 1)

Note that @parameters and @variables implicitly add () to values that are not given a call. The former specifies the values as known, while the latter specifies it as unknown. (..) signifies that the value should be left uncalled.

Sometimes it is convenient to define arrays of variables to model things like x₁,…,x₃. The @variables and @parameters macros support this with the following syntax:

@variables x[1:3];
x

3-element Array{Operation,1}:
 x₁()
 x₂()
 x₃()

# support for arbitrary ranges and tensors
@variables y[2:3,1:5:6];
y

2×2 Array{Operation,2}:
    y₂̒₁() y₂̒₆()
    y₃̒₁() y₃̒₆()

# also works for dependent variables
@parameters t; @variables z[1:3](t);
z

3-element Array{Operation,1}:
 z₁(t())
 z₂(t())
 z₃(t())

MomentClosure.ReactionSystemMod — Type

struct ReactionSystemMod

A system of chemical reactions. The structure is heavily based on ModelingToolkit.ReactionSystem and the relevant fields are left identical for the sake of familiar users.

Fields

iv

Independent variable (usually time)

states

Dependent (state) variables representing amount of each species.

ps

Parameter variables.

a

Propensity functions of all reactions

S

Stoichiometric matrix

Example

@parameters t, c₁, c₂, c₃, c₄, Ω
@variables X(t), Y(t)

# stoichiometric matrix
S_mat = [ 1 -1  1 -1;
         -1  1  0  0]

# propensity functions
a = [c₁*X*Y*(X-1)/Ω^2, c₂*X, c₃*Ω, c₄*X]

rn = ReactionSystemMod(t, [X, Y], [c₁, c₂, c₃, c₄, Ω], a, S_mat)

Basic model properties

To keep the API consistent with Catalyst, we provide a number of identical basic functions allowing easy access to network properties:

Catalyst.species — Function

species(rn::ReactionSystemMod)

Given a ReactionSystemMod, return a vector of species variables expressed as Term{Real}. Extension of Catalyst's species used with a ReactionSystem.

Catalyst.speciesmap — Function

speciesmap(rn::ReactionSystemMod)

Given a ReactionSystemMod, return a Dictionary mapping from species to species indices. Extension of Catalyst's speciesmap used with a ReactionSystem.

Catalyst.params — Function

params(rn::ReactionSystemMod)

Given a ReactionSystemMod, return a vector of parameter variables expressed as Sym{ModelingToolkit.Parameter{Real}. Extension of Catalyst's params used with a ReactionSystem.

Catalyst.paramsmap — Function

paramsmap(rn::ReactionSystemMod)

Given a ReactionSystemMod, return a Dictionary mapping from parameters to parameter indices. Extension of Catalyst's paramsmap used with a ReactionSystem.

Catalyst.numspecies — Function

numspecies(rn::ReactionSystemMod)

Return the number of species within the given ReactionSystemMod. Extension of Catalyst's numspecies used with a ReactionSystem.

Catalyst.numparams — Function

numparams(rn::ReactionSystemMod)

Return the number of parameters within the given ReactionSystemMod. Extension of Catalyst's numparams used with a ReactionSystem.

Catalyst.numreactions — Function

numreactions(rn::ReactionSystemMod)

Return the number of reactions within the given ReactionSystemMod. Extension of Catalyst's numreactions used with a ReactionSystem.

Moreover, we include two slight extensions of the Catalyst API:

MomentClosure.get_S_mat — Function

get_S_mat(rn::Union{ReactionSystem, ReactionSystemMod})

Return the (net) stoichiometric matrix of the given ReactionSystem or ReactionSystemMod.

MomentClosure.propensities — Function

propensities(rn::Union{ReactionSystem, ReactionSystemMod}; combinatoric_ratelaw=true)

Return a vector of propensity functions of all reactions in the given ReactionSystem or ReactionSystemMod.

Notes:

combinatoric_ratelaw=true uses binomials in calculating the propensity functions for a ReactionSystem, see the notes for ModelingToolkit.jumpratelaw. Note that this field is irrelevant using ReactionSystemMod as then the propensities are defined directly by the user.