MomentClosure.jl API

Model definition

As described in the tutorial on model initialisation, MomentClosure is fully compatible with reaction network models defined using Catalyst and stored as a ModelingToolkit.ReactionSystem. Nevertheless, we have implemented a limited extension of ModelingToolkit.ReactionSystem, called ReactionSystemMod, that allows us to consider systems containing reactions whose products can be independent geometrically distributed random variables. Note that such systems, one being the gene network example [LINK], are currently not supported by Catalyst. Moreover, ReactionSystemMod provides an alternative to Catalyst, as the model is now constructed directly in terms of user-defined reaction propensity functions and the stoichiometric matrix, with system parameters and species defined using ModelingToolkit @parameters and @variables respectively. However, we stress that whenever possible defining models using Catalyst is preferred due to its much better integration within the broader SciML ecosystem.

ModelingToolkit.@parameters — Macro

Define one or more known variables.

ModelingToolkit.@variables — Macro

Define one or more unknown variables.

@parameters t α σ(..) β[1:2]
@variables w(..) x(t) y() z(t, α, x)

expr = β₁* x + y^α + σ(3) * (z - t) - β₂ * w(t - 1)

Note that @parameters and @variables implicitly add () to values that are not given a call. The former specifies the values as known, while the latter specifies it as unknown. (..) signifies that the value should be left uncalled.

Sometimes it is convenient to define arrays of variables to model things like x₁,…,x₃. The @variables and @parameters macros support this with the following syntax:

@variables x[1:3];
x

3-element Array{Operation,1}:
 x₁()
 x₂()
 x₃()

# support for arbitrary ranges and tensors
@variables y[2:3,1:5:6];
y

2×2 Array{Operation,2}:
    y₂̒₁() y₂̒₆()
    y₃̒₁() y₃̒₆()

# also works for dependent variables
@parameters t; @variables z[1:3](t);
z

3-element Array{Operation,1}:
 z₁(t())
 z₂(t())
 z₃(t())

MomentClosure.ReactionSystemMod — Type

struct ReactionSystemMod

A system of chemical reactions. The structure is heavily based on ModelingToolkit.ReactionSystem and the relevant fields are left identical for the sake of familiar users.

Fields

iv

Independent variable (usually time)

states

Dependent (state) variables representing amount of each species.

ps

Parameter variables.

a

Propensity functions of all reactions

S

Stoichiometric matrix

Example

@parameters t, c₁, c₂, c₃, c₄, Ω
@variables X(t), Y(t)

# stoichiometric matrix
S_mat = [ 1 -1  1 -1;
         -1  1  0  0]

# propensity functions
a = [c₁*X*Y*(X-1)/Ω^2, c₂*X, c₃*Ω, c₄*X]

rn = ReactionSystemMod(t, [X, Y], [c₁, c₂, c₃, c₄, Ω], a, S_mat)