MomentClosure.jl API

Model definition

MomentClosure is fully compatible with reaction network models defined using Catalyst and stored as a ModelingToolkit.ReactionSystem. Nevertheless, we have implemented a limited extension of ModelingToolkit.ReactionSystem, called ReactionSystemMod, that allows us to consider systems containing reactions whose products can be independent geometrically distributed random variables. For example, such reactions can be encountered in certain models of autoregulatory gene networks. Moreover, ReactionSystemMod provides an alternative to Catalyst as the model is now constructed directly in terms of user-defined reaction propensity functions and the stoichiometric matrix, with system parameters and species defined using ModelingToolkit @parameters and @variables respectively. However, we stress that defining models using Catalyst is preferred whenever possible due to its much richer integration within the broader SciML ecosystem.

ModelingToolkit.@parameters — Macro

Define one or more known variables.

Missing docstring.

Missing docstring for ModelingToolkit.@variables. Check Documenter's build log for details.

MomentClosure.ReactionSystemMod — Type

struct ReactionSystemMod

A system of chemical reactions. The structure is heavily based on ModelingToolkit.ReactionSystem and the relevant fields are left identical for the sake of familiar users.

Fields

iv

Independent variable (usually time)

states

Dependent (state) variables representing amount of each species.

ps

Parameter variables.

a

Propensity functions of all reactions

S

Stoichiometric matrix

Example

@parameters t, c₁, c₂, c₃, c₄, Ω
@variables X(t), Y(t)

# stoichiometric matrix
S_mat = [ 1 -1  1 -1;
         -1  1  0  0]

# propensity functions
a = [c₁*X*Y*(X-1)/Ω^2, c₂*X, c₃*Ω, c₄*X]

rn = ReactionSystemMod(t, [X, Y], [c₁, c₂, c₃, c₄, Ω], a, S_mat)

Basic model properties

To keep the API consistent with Catalyst, we provide a number of identical basic functions allowing easy access to network properties:

Catalyst.species — Function

species(rn::ReactionSystemMod)

Given a ReactionSystemMod, return a vector of species variables expressed as Term{Real}. Extension of Catalyst's species used with a ReactionSystem.

Catalyst.speciesmap — Function

speciesmap(rn::ReactionSystemMod)

Given a ReactionSystemMod, return a Dictionary mapping from species to species indices. Extension of Catalyst's speciesmap used with a ReactionSystem.

Catalyst.params — Function

params(rn::ReactionSystemMod)

Given a ReactionSystemMod, return a vector of parameter variables expressed as Sym{ModelingToolkit.Parameter{Real}. Extension of Catalyst's params used with a ReactionSystem.

Catalyst.paramsmap — Function

paramsmap(rn::ReactionSystemMod)

Given a ReactionSystemMod, return a Dictionary mapping from parameters to parameter indices. Extension of Catalyst's paramsmap used with a ReactionSystem.

Catalyst.numspecies — Function

numspecies(rn::ReactionSystemMod)

Return the number of species within the given ReactionSystemMod. Extension of Catalyst's numspecies used with a ReactionSystem.

Catalyst.numparams — Function

numparams(rn::ReactionSystemMod)

Return the number of parameters within the given ReactionSystemMod. Extension of Catalyst's numparams used with a ReactionSystem.

Catalyst.numreactions — Function

numreactions(rn::ReactionSystemMod)

Return the number of reactions within the given ReactionSystemMod. Extension of Catalyst's numreactions used with a ReactionSystem.

Moreover, we include two slight extensions of the Catalyst API:

MomentClosure.get_S_mat — Function

get_S_mat(rn::Union{ReactionSystem, ReactionSystemMod})

Return the (net) stoichiometric matrix of the given ReactionSystem or ReactionSystemMod.

MomentClosure.propensities — Function

propensities(rn::Union{ReactionSystem, ReactionSystemMod}; combinatoric_ratelaw=true)

Return a vector of propensity functions of all reactions in the given ReactionSystem or ReactionSystemMod.

Notes:

combinatoric_ratelaw=true uses binomials in calculating the propensity functions of a ReactionSystem, see the notes for ModelingToolkit.jumpratelaw. Note that this field is irrelevant using ReactionSystemMod as then the propensities are defined directly by the user.

Moment Equations

MomentClosure.generate_raw_moment_eqs — Function

generate_raw_moment_eqs(rn::Union{ReactionSystem, ReactionSystemMod}, m_order::Int; combinatoric_ratelaw = true)

Given a ReactionSystem or ReactionSystemMod, return the RawMomentEquations of the system generated up to m_order.

Notes:

The expansion order $q$, denoted by q_order throughout the docs, is automatically determined from the given polynomial form of the propensity functions, see the tutorial and the theory section for more details on how q_order is obtained.

combinatoric_ratelaw=true uses binomials in calculating the propensity functions of a ReactionSystem, see the notes for ModelingToolkit.jumpratelaw. Note that this field is irrelevant using ReactionSystemMod as then the propensities are defined directly by the user.

MomentClosure.RawMomentEquations — Type

struct RawMomentEquations <: MomentClosure.MomentEquations

Raw moment equations generated for the given system plus a number of helper parameters (important for internal functionality).

Fields

odes

ModelingToolkit.ODESystem containing the time-evolution equations of raw moments.

μ

Symbolic variables defining the raw moments.

N

Number of species within the system.

m_order

Order of moment equations.

q_order

Expansion order.

iter_all

Vector of all index combinations (Tuples) up to q_order.

iter_m

Vector of all index combinations (Tuples) up to m_order.

iter_q

Vector of all index combinations (Tuples) of order greater than m_order up to q_order.

iter_1

Vector of index combinations (Tuples)of order 1.

MomentClosure.generate_central_moment_eqs — Function

generate_central_moment_eqs(rn::Union{ReactionSystem, ReactionSystemMod}, m_order::Int, q_order=nothing; combinatoric_ratelaw = true)

Given a ReactionSystem or ReactionSystemMod, return the CentralMomentEquations of the system generated up to m_order.

Notes:

if q_order is not specified by the user, it is assumed that the reaction network contains only polynomial propensity functions and hence q_order is determined automatically as in generate_raw_moment_eqs. However, q_order must be specified if non-polynomial propensities are included. Note that the expansion order $q$ denotes the highest order of central moments which will be included in the ODEs (due to the Taylor expansion of propensity functions).

combinatoric_ratelaw=true uses binomials in calculating the propensity functions of a ReactionSystem, see the notes for ModelingToolkit.jumpratelaw. Note that this field is irrelevant using ReactionSystemMod as then the propensities are defined directly by the user.

MomentClosure.CentralMomentEquations — Type

struct CentralMomentEquations <: MomentClosure.MomentEquations

Central moment equations generated for the given system plus a number of helper parameters (important for internal functionality).

Fields

odes

Moment ODEs describing the evolution of central moments.

μ

Symbolic variables defining the means.

M

Symbolic variables defining the central moments.

N

Number of species within the system.

m_order

Order of moment equations.

q_order

Expansion order.

iter_all

Vector of all index combinations up to q_order.

iter_m

Vector of all index combinations up to m_order.

iter_q

Vector of all index combinations of order greater than m_order up to q_order.

iter_1

Vector of index combinations of order 1.

Moment Closure

Displaying equations and closures

MomentClosure.format_moment_eqs — Function

format_moment_eqs(odes::ODESystem)

Given a ModelingToolkit.ODESystem containing the moment equations, return an array of formatted strings representing the ODEs that can be visualised using Latexify. Although Latexify can be applied directly on the ODESystem, format_moment_eqs makes it visually more appealing: simplifies the symbolic expressions further (expanding all terms), makes the time-dependence of all moment variables implicit (removes all (t)) and removes all trailing zeros (2.0 → 2).

Missing docstring.

Missing docstring for format_closure. Check Documenter's build log for details.