MomentClosure.jl API

Model definition

MomentClosure is fully compatible with reaction network models defined using Catalyst and stored as a ModelingToolkit.ReactionSystem. Nevertheless, we have implemented a limited extension of ModelingToolkit.ReactionSystem, called ReactionSystemMod, that allows us to consider systems containing reactions whose products can be independent geometrically distributed random variables. For example, such reactions can be encountered in certain models of autoregulatory gene networks. Moreover, ReactionSystemMod provides an alternative to Catalyst as the model is now constructed directly in terms of user-defined reaction propensity functions and the stoichiometric matrix, with system parameters and species defined using ModelingToolkit @parameters and Symbolics @variables respectively. However, we stress that defining models using Catalyst is preferred whenever possible due to its much richer integration within the broader SciML ecosystem.

ModelingToolkit.@parameters — Macro

Define one or more known variables.

Symbolics.@variables — Macro

Define one or more unknown variables.

@parameters t α σ(..) β[1:2]
@variables w(..) x(t) y z(t, α, x)

expr = β₁* x + y^α + σ(3) * (z - t) - β₂ * w(t - 1)

(..) signifies that the value should be left uncalled.

Sometimes it is convenient to define arrays of variables to model things like x₁,…,x₃. The @variables macro supports this with the following syntax:

@variables x[1:3];
x

3-element Vector{Num}:
 x₁
 x₂
 x₃

# support for arbitrary ranges and tensors
@variables y[2:3,1:5:6];
y

2×2 Matrix{Num}:
 y₂ˏ₁  y₂ˏ₆
 y₃ˏ₁  y₃ˏ₆

# also works for dependent variables
@variables t z[1:3](t);
z

3-element Array{Num,1}:
 z₁(t)
 z₂(t)
 z₃(t)

MomentClosure.ReactionSystemMod — Type

struct ReactionSystemMod

A system of chemical reactions. The structure is heavily based on ModelingToolkit.ReactionSystem and the relevant fields are left identical for the sake of familiar users.

Fields

iv

Independent variable (usually time)

states

Dependent (state) variables representing amount of each species.

ps

Parameter variables.

a

Propensity functions of all reactions

S

Stoichiometric matrix

Example

@parameters t, c₁, c₂, c₃, c₄, Ω
@variables X(t), Y(t)

# stoichiometric matrix
S_mat = [ 1 -1  1 -1;
         -1  1  0  0]

# propensity functions
a = [c₁*X*Y*(X-1)/Ω^2, c₂*X, c₃*Ω, c₄*X]

rn = ReactionSystemMod(t, [X, Y], [c₁, c₂, c₃, c₄, Ω], a, S_mat)

Basic model properties

To keep the API consistent with Catalyst, we provide a number of identical basic functions allowing easy access to network properties:

Catalyst.species — Function

species(rn::ReactionSystemMod)

Given a ReactionSystemMod, return a vector of species variables expressed as Term{Real}. Extension of Catalyst's species used with a ReactionSystem.

Catalyst.speciesmap — Function

speciesmap(rn::ReactionSystemMod)

Given a ReactionSystemMod, return a Dictionary mapping from species to species indices. Extension of Catalyst's speciesmap used with a ReactionSystem.

Catalyst.params — Function

params(rn::ReactionSystemMod)

Given a ReactionSystemMod, return a vector of parameter variables expressed as Sym{ModelingToolkit.Parameter{Real}. Extension of Catalyst's params used with a ReactionSystem.

Catalyst.paramsmap — Function

paramsmap(rn::ReactionSystemMod)

Given a ReactionSystemMod, return a Dictionary mapping from parameters to parameter indices. Extension of Catalyst's paramsmap used with a ReactionSystem.

Catalyst.numspecies — Function

numspecies(rn::ReactionSystemMod)

Return the number of species within the given ReactionSystemMod. Extension of Catalyst's numspecies used with a ReactionSystem.

Catalyst.numparams — Function

numparams(rn::ReactionSystemMod)

Return the number of parameters within the given ReactionSystemMod. Extension of Catalyst's numparams used with a ReactionSystem.

Catalyst.numreactions — Function

numreactions(rn::ReactionSystemMod)

Return the number of reactions within the given ReactionSystemMod. Extension of Catalyst's numreactions used with a ReactionSystem.

Moreover, we include two slight extensions of the Catalyst API:

MomentClosure.get_S_mat — Function

get_S_mat(rn::Union{ReactionSystem, ReactionSystemMod})

Return the (net) stoichiometric matrix of the given ReactionSystem or ReactionSystemMod.

MomentClosure.propensities — Function

propensities(rn::Union{ReactionSystem, ReactionSystemMod}; combinatoric_ratelaw=true)

Return a vector of propensity functions of all reactions in the given ReactionSystem or ReactionSystemMod.

Notes:

combinatoric_ratelaw=true uses binomials in calculating the propensity functions of a ReactionSystem, see the notes for ModelingToolkit.jumpratelaw. Note that this field is irrelevant using ReactionSystemMod as then the propensities are defined directly by the user.

Moment Equations

MomentClosure.generate_raw_moment_eqs — Function

generate_raw_moment_eqs(rn::Union{ReactionSystem, ReactionSystemMod}, m_order::Int; combinatoric_ratelaw = true)

Given a ReactionSystem or ReactionSystemMod, return the RawMomentEquations of the system generated up to m_order.

Notes:

The expansion order $q$, denoted by q_order throughout the docs, is automatically determined from the given polynomial form of the propensity functions, see the tutorial and the theory section for more details on how q_order is obtained.

combinatoric_ratelaw=true uses binomials in calculating the propensity functions of a ReactionSystem, see the notes for ModelingToolkit.jumpratelaw. Note that this field is irrelevant using ReactionSystemMod as then the propensities are defined directly by the user.

MomentClosure.RawMomentEquations — Type

struct RawMomentEquations <: MomentClosure.MomentEquations

Raw moment equations generated for the given system plus a number of helper parameters (important for internal functionality).

Fields

odes

ModelingToolkit.ODESystem consisting of the time-evolution equations of raw moments.

μ

Symbolic variables defining the raw moments.

N

Number of species within the system.

m_order

Order of moment equations.

q_order

Expansion order.

iter_all

Vector of all index combinations (Tuples) up to q_order.

iter_m

Vector of all index combinations (Tuples) up to m_order.

iter_q

Vector of all index combinations (Tuples) of order greater than m_order up to q_order.

iter_1

Vector of index combinations (Tuples)of order 1.

MomentClosure.generate_central_moment_eqs — Function

generate_central_moment_eqs(rn::Union{ReactionSystem, ReactionSystemMod}, m_order::Int, q_order=nothing; combinatoric_ratelaw = true)

Given a ReactionSystem or ReactionSystemMod, return the CentralMomentEquations of the system generated up to m_order.

Notes:

if q_order is not specified by the user, it is assumed that the reaction network contains only polynomial propensity functions and hence q_order is determined automatically as in generate_raw_moment_eqs. However, q_order must be specified if non-polynomial propensities are included. Note that the expansion order $q$ denotes the highest order of central moments which will be included in the ODEs (due to the Taylor expansion of propensity functions).

combinatoric_ratelaw=true uses binomials in calculating the propensity functions of a ReactionSystem, see the notes for ModelingToolkit.jumpratelaw. Note that this field is irrelevant using ReactionSystemMod as then the propensities are defined directly by the user.

MomentClosure.CentralMomentEquations — Type

struct CentralMomentEquations <: MomentClosure.MomentEquations

Central moment equations generated for the given system plus a number of helper parameters (important for internal functionality).

Fields

odes

ModelingToolkit.ODESystem consisting of the time-evolution equations of central moments.

μ

Symbolic variables defining the means.

M

Symbolic variables defining the central moments.

N

Number of species within the system.

m_order

Order of moment equations.

q_order

Expansion order.

iter_all

Vector of all index combinations up to q_order.

iter_m

Vector of all index combinations up to m_order.

iter_q

Vector of all index combinations of order greater than m_order up to q_order.

iter_1

Vector of index combinations of order 1.

Moment Closure

MomentClosure.moment_closure — Function

moment_closure(sys::MomentEquations, closure::String, binary_vars::Vector=[]; clean=true)

Given MomentEquations, apply the specified moment closure approximation and return the ClosedMomentEquations.

The supported closure options are:

"zero"

"normal"

"log-normal"

"poisson"

"gamma"

"derivative matching"

"conditional gaussian"

"conditional derivative matching"

Notes

binary_vars is required in case of conditional closures. It must be specified by denoting in an array the indices of all species whose molecule number is a Bernoulli (binary) variable.

clean argument is only relevant for gamma closure. Namely, clean=true implies that the closure functions and the resulting closed moment equations will be heavily simplified in order to preserve numerical stability. However, this can extremely computationally expensive due to complicated functional form of the higher order moments under gamma closure. Setting clean=false is generally much faster as the symbolic expressions are not simplified internally, but it can lead to severe numerical instabilities.

MomentClosure.ClosedMomentEquations — Type

struct ClosedMomentEquations <: MomentClosure.MomentEquations

Closed moment equations and the corresponding closure functions.

Fields

odes

ModelingToolkit.ODESystem consisting of the time-evolution equations of closed moments.

closure

Dictionary of moment closure functions for each higher order moment.

open_eqs

Original raw or central moment equations (before closure was applied).

Missing docstring.

Missing docstring for deterministic_IC. Check Documenter's build log for details.

Displaying Equations and Closures

The generated moment equations can be converted into LaTeX expressions using Latexify as:

using Latexify
latexify(moment_eqs)

Note that the ModelingToolkit.ODESystem (saved as moment_eqs.odes) can also be passed to latexify function directly.

Given ClosedMomentEquations, the closure functions can be visualised in the same way by adding :closure argument:

latexify(moment_eqs, :closure)

Stochastic Simulation Utilities

Missing docstring.

Missing docstring for sample_raw_moments. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sample_central_moments. Check Documenter's build log for details.

Missing docstring.

Missing docstring for sample_cumulants. Check Documenter's build log for details.