MomentClosure.jl API

Model definition

MomentClosure is fully compatible with reaction network models defined using Catalyst and stored as a ModelingToolkit.ReactionSystem. Nevertheless, we have implemented our own ReactionSystemMod, that allows us to consider systems containing reactions which products are independent geometrically distributed random variables (currently unsupported by Catalyst). For example, such reactions can be encountered in certain types of gene regulatory network models, as discussed in this tutorial. Moreover, ReactionSystemMod provides an alternative to Catalyst as the model can be now constructed directly in terms of user-defined reaction propensity functions and the stoichiometric matrix, with system parameters and species defined using ModelingToolkit @parameters and Symbolics @variables respectively.

We stress that ReactionSystemMod, as the name suggests, is a rather trivial extension/modification of ModelingToolkit.ReactionSystem. Although ReactionSystemMod provides the same basic functions to access network properties as does Catalyst for a ModelingToolkit.ReactionSystem (see below), its integration within the broader SciML framework is very limited (only compatible with SSA solvers) and so we advise using Catalyst for model initialisation whenever possible. Finally, note that ReactionSystemMod is a temporary solution: we expect this interface to be supported by Catalyst directly soon (see the discussion here and here).

ModelingToolkit.@parameters — Macro

Define one or more known variables.

Symbolics.@variables — Macro

Define one or more unknown variables.

@variables t α σ(..) β[1:2]
@variables w(..) x(t) y z(t, α, x)

expr = β₁* x + y^α + σ(3) * (z - t) - β₂ * w(t - 1)

(..) signifies that the value should be left uncalled.

Sometimes it is convenient to define arrays of variables to model things like x₁,…,x₃. The @variables macro supports this with the following syntax:

@variables x[1:3];
x

3-element Vector{Num}:
 x₁
 x₂
 x₃

# support for arbitrary ranges and tensors
@variables y[2:3,1:5:6];
y

2×2 Matrix{Num}:
 y₂ˏ₁  y₂ˏ₆
 y₃ˏ₁  y₃ˏ₆

# also works for dependent variables
@variables t z[1:3](t);
z

3-element Array{Num,1}:
 z₁(t)
 z₂(t)
 z₃(t)

MomentClosure.ReactionSystemMod — Type

struct ReactionSystemMod

A system of chemical reactions. The structure is heavily based on ModelingToolkit.ReactionSystem and the relevant fields are left identical for the sake of familiar users.

Fields

iv

Independent variable (usually time)

states

Dependent (state) variables representing amount of each species.

ps

Parameter variables.

a

Propensity functions of all reactions

S

Stoichiometric matrix

Notes

The molecule numbers of each chemical species must be explicitly defined as Symbolics.@variables and reaction constants as ModelingToolkit.@parameters. The time, $t$, must also be initialised as a ModelingToolkit.parameter and be used to indicate the time-dependence of species' variables.

To specify that a certain reaction product is geometrically distributed, the corresponding stoichiometric matrix element must be a ModelingToolkit.parameter denoting the mean of the distribution.

Example (also see here)

using MomentClosure

@parameters t, c₁, c₂, c₃, c₄, Ω
@variables X(t), Y(t)

# stoichiometric matrix
S_mat = [ 1 -1  1 -1;
         -1  1  0  0]

# propensity functions
a = [c₁*X*Y*(X-1)/Ω^2, c₂*X, c₃*Ω, c₄*X]

rn = ReactionSystemMod(t, [X, Y], [c₁, c₂, c₃, c₄, Ω], a, S_mat)

Basic model properties

To keep the API consistent with Catalyst, we provide a number of identical basic functions allowing easy access to network properties:

Catalyst.species — Function

species(rn::ReactionSystemMod)

Given a ReactionSystemMod, return a vector of species variables expressed as Term{Real}. Extension of Catalyst's species used with a ReactionSystem.

Catalyst.speciesmap — Function

speciesmap(rn::ReactionSystemMod)

Given a ReactionSystemMod, return a Dictionary mapping from species to species indices. Extension of Catalyst's speciesmap used with a ReactionSystem.

Catalyst.params — Function

params(rn::ReactionSystemMod)

Given a ReactionSystemMod, return a vector of parameter variables expressed as Sym{ModelingToolkit.Parameter{Real}. Extension of Catalyst's params used with a ReactionSystem.

Catalyst.paramsmap — Function

paramsmap(rn::ReactionSystemMod)

Given a ReactionSystemMod, return a Dictionary mapping from parameters to parameter indices. Extension of Catalyst's paramsmap used with a ReactionSystem.

Catalyst.numspecies — Function

numspecies(rn::ReactionSystemMod)

Return the number of species within the given ReactionSystemMod. Extension of Catalyst's numspecies used with a ReactionSystem.

Catalyst.numparams — Function

numparams(rn::ReactionSystemMod)

Return the number of parameters within the given ReactionSystemMod. Extension of Catalyst's numparams used with a ReactionSystem.

Catalyst.numreactions — Function

numreactions(rn::ReactionSystemMod)

Return the number of reactions within the given ReactionSystemMod. Extension of Catalyst's numreactions used with a ReactionSystem.

Moreover, we include two slight extensions of the Catalyst API:

MomentClosure.get_S_mat — Function

get_S_mat(rn::Union{ReactionSystem, ReactionSystemMod})

Return the (net) stoichiometric matrix of the given ReactionSystem or ReactionSystemMod.

MomentClosure.propensities — Function

propensities(rn::Union{ReactionSystem, ReactionSystemMod}; combinatoric_ratelaw=true)

Return a vector of propensity functions of all reactions in the given ReactionSystem or ReactionSystemMod.

Notes:

combinatoric_ratelaw=true uses binomials in calculating the propensity functions of a ReactionSystem, see the notes for ModelingToolkit.jumpratelaw. Note that this field is irrelevant using ReactionSystemMod as then the propensities are defined directly by the user.

Moment Equations

MomentClosure.generate_raw_moment_eqs — Function

generate_raw_moment_eqs(rn::Union{ReactionSystem, ReactionSystemMod}, m_order::Int; combinatoric_ratelaw = true)

Given a ReactionSystem or ReactionSystemMod, return the RawMomentEquations of the system generated up to m_order.

Notes:

The expansion order $q$, denoted by q_order throughout the docs, is automatically determined from the given polynomial form of the propensity functions, see the tutorial and the theory section for more details on how q_order is obtained.

combinatoric_ratelaw=true uses binomials in calculating the propensity functions of a ReactionSystem, see the notes for ModelingToolkit.jumpratelaw. Note that this field is irrelevant using ReactionSystemMod as then the propensities are defined directly by the user.

MomentClosure.RawMomentEquations — Type

struct RawMomentEquations <: MomentClosure.MomentEquations

Raw moment equations generated for the given system plus a number of helper parameters (used internally).

Fields

odes

ModelingToolkit.ODESystem consisting of the time-evolution equations of raw moments.

μ

Symbolic variables defining the raw moments.

N

Number of species within the system.

m_order

Order of moment equations.

q_order

Expansion order.

iter_all

Vector of all index combinations up to q_order.

iter_m

Vector of all index combinations up to m_order.

iter_q

Vector of all index combinations of order greater than m_order up to q_order.

iter_1

Vector of index combinations of order 1.

MomentClosure.generate_central_moment_eqs — Function

generate_central_moment_eqs(rn::Union{ReactionSystem, ReactionSystemMod}, m_order::Int, q_order=nothing; combinatoric_ratelaw = true)

Given a ReactionSystem or ReactionSystemMod, return the CentralMomentEquations of the system generated up to m_order.

Notes:

if q_order is not specified by the user, it is assumed that the reaction network contains only polynomial propensity functions and hence q_order is determined automatically as in generate_raw_moment_eqs. However, q_order must be specified if non-polynomial propensities are included. Note that the expansion order $q$ denotes the highest order of central moments which will be included in the ODEs (due to the Taylor expansion of propensity functions).

combinatoric_ratelaw=true uses binomials in calculating the propensity functions of a ReactionSystem, see the notes for ModelingToolkit.jumpratelaw. Note that this field is irrelevant using ReactionSystemMod as then the propensities are defined directly by the user.

MomentClosure.CentralMomentEquations — Type

struct CentralMomentEquations <: MomentClosure.MomentEquations

Central moment equations generated for the given system plus a number of helper parameters (used internally).

Fields

odes

ModelingToolkit.ODESystem consisting of the time-evolution equations of central moments.

μ

Symbolic variables defining the means.

M

Symbolic variables defining the central moments.

N

Number of species within the system.

m_order

Order of moment equations.

q_order

Expansion order.

iter_all

Vector of all index combinations up to q_order.

iter_m

Vector of all index combinations up to m_order.

iter_q

Vector of all index combinations of order greater than m_order up to q_order.

iter_1

Vector of index combinations of order 1.

MomentClosure.bernoulli_moment_eqs — Function

bernoulli_moment_eqs(sys::MomentEquations, binary_vars::Array{Int,1})

Given MomentEquations and an array of indices specifying the species which molecule numbers are binary variables (either 0 or 1), apply identities of Bernoulli variables to remove the redundant ODEs and return the cleaned up MomentEquations. See here for example usage.

Moment Closure

MomentClosure.moment_closure — Function

moment_closure(sys::MomentEquations, closure::String, binary_vars::Array{Int,1}=Int[])

Given MomentEquations, apply the specified moment closure approximation and return the ClosedMomentEquations.

The supported closure options are:

"zero"

"normal"

"log-normal"

"poisson"

"gamma"

"derivative matching"

"conditional gaussian"

"conditional derivative matching"

Notes

binary_vars must be specified for conditional closures as an array of indices of all species (as in speciesmap) which molecule number is a Bernoulli variable. Although not necessary for other closures, specifying binary_vars is recommended as the properties of Bernoulli variables will be used to remove the redundant moment equations and simplify the expressions, which can significantly improve numerical stability.

MomentClosure.ClosedMomentEquations — Type

struct ClosedMomentEquations <: MomentClosure.MomentEquations

Closed moment equations and the corresponding closure functions.

Fields

odes

ModelingToolkit.ODESystem consisting of the time-evolution equations of closed moments.

closure

Dictionary of moment closure functions for each higher order moment.

open_eqs

Original raw or central moment equations (before closure was applied).

MomentClosure.deterministic_IC — Function

deterministic_IC(u₀::Array{T, 1}, eqs::MomentEquations) where T<:Real

Given an array of initial molecule numbers and the corresponding moment equations, return a mapping of each moment to its initial value under deterministic initial conditions.

Notes

The means are set to initial molecule numbers (as they take the values specified in u₀ with probability one). The higher order raw moments are products of the corresponding powers of the means whereas the higher order central moments are simply zero.

The ordering of u₀ elements must be consistent with the ordering of species in the corresponding reaction system (can be checked with the speciesmap function).

As higher-order moment functions under log-normal, gamma, derivative matching and the conditional closures involve moments raised to negative powers, setting initial molecule numbers of certain species to zeros will result in NaN errors when solving the ODEs (the specifics depend on the system at hand).

Displaying Equations and Closures

The generated moment equations can be converted into LaTeX expressions using Latexify as:

using Latexify
latexify(moment_eqs)

A ModelingToolkit.ODESystem (saved as moment_eqs.odes) can also be passed to latexify function directly but the output will be different as we apply additional formatting to the symbolic expressions.

Given ClosedMomentEquations, the closure functions can be visualised in the same way by adding a :closure argument:

latexify(moment_eqs, :closure)

Note that this will print out only those higher order moments which are found in the given moment equations. It is possible to print the closure functions of all higher order moments using print_all=true argument:

latexify(moment_eqs, :closure, print_all=true)

Stochastic Simulation Utilities

A ReactionSystemMod is compatible with the standard DifferentialEquations JumpProblem and hence can be simulated using SSA. Internally, ReactionSystemMod is converted into a system of ConstantRateJumps taking into account the geometrically distributed reaction products. Note, however, that the implementation here is very restricted: time-dependent propensities are currently not allowed and MassActionJumps cannot be used due to limitations of the ReactionSystemMod API, hence the performance is subpar.

We also provide provides functions for higher-order moment extraction from SSA data:

MomentClosure.get_raw_moments — Function

get_raw_moments(sol::EnsembleSolution, order::Int; naive::Bool=true, b::Int=2)

Given an EnsembleSolution of DifferentialEquations ensemble simulation, return a Dictionary of raw moments computed up to the specified order at each time step.

Notes

For example, the dictionary key (2,0,1) maps to an array containing the values of the raw moment $μ_{201}$ at each time step.

It is assumed that the time steps are all at the same time point for all trajectories (i.e., fixed dt used by the integrator or values were saved using saveat, as discussed here).

Moments are computed using Cumulants.jl internally. The naive algorithm of moment tensor calculations (naive=true) is usually faster for small systems but the proposed novel algorithm (naive=false) should be more efficient in case of many marginal variables. The block size b can also be specified for the novel algorithm and may have a significant effect on its performance.

Only useful if higher order moments are needed: DifferentialEquations has a number of far more efficient and flexible ensemble statistics functions for means, variances and correlations, see this tutorial for more details.

MomentClosure.get_central_moments — Function

get_central_moments(sol::EnsembleSolution, order::Int; naive::Bool=true, b::Int=2)

Given an EnsembleSolution of DifferentialEquations ensemble simulation, return a Dictionary of central moment estimates computed up to the specified order at each time step. See the notes of get_raw_moments function for more information.

MomentClosure.get_cumulants — Function

get_cumulants(sol::EnsembleSolution, order::Int; naive::Bool=true, b::Int=2)

Given an EnsembleSolution of DifferentialEquations ensemble simulation, return a Dictionary of cumulant estimates computed up to the specified order at each time step. See the notes of get_raw_moments function for more information.